+++ homematicip/aio/class_maps.py
@@ -6,6 +6,7 @@ from homematicip.aio.securityEvent import *
 
 TYPE_CLASS_MAP = {
     DeviceType.DEVICE: AsyncDevice,
+    DeviceType.ACCELERATION_SENSOR: AsyncAccelerationSensor,
     DeviceType.ALARM_SIREN_INDOOR: AsyncAlarmSirenIndoor,
     DeviceType.BRAND_BLIND: AsyncBrandBlind,
     DeviceType.BRAND_DIMMER: AsyncBrandDimmer,