+++ homematicip_demo/fake_cloud_server.py
@@ -266,6 +266,8 @@ class FakeCloudServer:
 
         return response
 
+    # endregion
+
     # region rule
     @validate_authorization
     def post_hmip_rule_enableSimpleRule(self, request: Request, response: Response):
@@ -555,6 +557,95 @@ class FakeCloudServer:
             "simpleRGBColorState"
         ]
         # not sure what to do with onTime and rampTime :/
+        response.status_code = 200
+        return response
+
+    @validate_authorization
+    def post_hmip_device_configuration_setNotificationSoundType(
+        self, request: Request, response: Response
+    ):
+
+        js = json.loads(request.data)
+        d = self.data["devices"][js["deviceId"]]
+        channelIndex = "{}".format(js["channelIndex"])
+        notificationSoundTypeKey = (
+            "notificationSoundTypeHighToLow"
+            if js["isHighToLow"]
+            else "notificationSoundTypeLowToHigh"
+        )
+        d["functionalChannels"][channelIndex][notificationSoundTypeKey] = js[
+            "notificationSoundType"
+        ]
+        response.status_code = 200
+        return response
+
+    @validate_authorization
+    def post_hmip_device_configuration_setAccelerationSensorEventFilterPeriod(
+        self, request: Request, response: Response
+    ):
+
+        js = json.loads(request.data)
+        d = self.data["devices"][js["deviceId"]]
+        channelIndex = "{}".format(js["channelIndex"])
+        d["functionalChannels"][channelIndex][
+            "accelerationSensorEventFilterPeriod"
+        ] = js["accelerationSensorEventFilterPeriod"]
+        response.status_code = 200
+        return response
+
+    @validate_authorization
+    def post_hmip_device_configuration_setAccelerationSensorTriggerAngle(
+        self, request: Request, response: Response
+    ):
+
+        js = json.loads(request.data)
+        d = self.data["devices"][js["deviceId"]]
+        channelIndex = "{}".format(js["channelIndex"])
+        d["functionalChannels"][channelIndex]["accelerationSensorTriggerAngle"] = js[
+            "accelerationSensorTriggerAngle"
+        ]
+        response.status_code = 200
+        return response
+
+    @validate_authorization
+    def post_hmip_device_configuration_setAccelerationSensorSensitivity(
+        self, request: Request, response: Response
+    ):
+
+        js = json.loads(request.data)
+        d = self.data["devices"][js["deviceId"]]
+        channelIndex = "{}".format(js["channelIndex"])
+        d["functionalChannels"][channelIndex]["accelerationSensorSensitivity"] = js[
+            "accelerationSensorSensitivity"
+        ]
+        response.status_code = 200
+        return response
+
+    @validate_authorization
+    def post_hmip_device_configuration_setAccelerationSensorNeutralPosition(
+        self, request: Request, response: Response
+    ):
+
+        js = json.loads(request.data)
+        d = self.data["devices"][js["deviceId"]]
+        channelIndex = "{}".format(js["channelIndex"])
+        d["functionalChannels"][channelIndex]["accelerationSensorNeutralPosition"] = js[
+            "accelerationSensorNeutralPosition"
+        ]
+        response.status_code = 200
+        return response
+
+    @validate_authorization
+    def post_hmip_device_configuration_setAccelerationSensorMode(
+        self, request: Request, response: Response
+    ):
+
+        js = json.loads(request.data)
+        d = self.data["devices"][js["deviceId"]]
+        channelIndex = "{}".format(js["channelIndex"])
+        d["functionalChannels"][channelIndex]["accelerationSensorMode"] = js[
+            "accelerationSensorMode"
+        ]
         response.status_code = 200
         return response