+++ CHANGELOG.md
@@ -8,7 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Devices
   - HMIP-MOD-RC8 (Open Collector Module Sender - 8x)
+  - HMIP-SAM        (Acceleration Sensor)
 
+### Deprecated
+- API
+  - moved homematicip.HomematicIPBaseObject to homematicip.base.HomematicIPBaseObject package 
 ## [0.10.10] - 2019-08-01
 ### Added
 - Devices